+++ tests/Value/StreetNameTest.php
@@ -12,27 +12,45 @@ use DigipolisGent\Flanders\BasicRegisters\Value\StreetNameId;
 use PHPUnit\Framework\TestCase;
 
 /**
+ * @covers \DigipolisGent\Flanders\BasicRegisters\Value\AbstractWithGeographicalNames
  * @covers \DigipolisGent\Flanders\BasicRegisters\Value\StreetName
  */
 class StreetNameTest extends TestCase
 {
 
     /**
-     * Street name is created from its object id and geographical names.
+     * Value is created from its details.
      *
      * @test
      */
-    public function createdFromObjectIdAndGeographicalNames(): void
+    public function valueIsCreatedFromItsDetails(): void
     {
         $streetNameId = new StreetNameId(123);
         $geographicalNames = new GeographicalNames(
             new GeographicalName(new LanguageCode('NL'), 'Foo Nl')
         );
 
-        $locality = new StreetName($streetNameId, $geographicalNames);
+        $streetName = new StreetName($streetNameId, $geographicalNames);
 
-        $this->assertSame($streetNameId, $locality->streetNameId());
-        $this->assertSame($geographicalNames, $locality->geographicalNames());
+        $this->assertSame($streetNameId, $streetName->streetNameId());
+        $this->assertSame($geographicalNames, $streetName->geographicalNames());
+    }
+
+    /**
+     * The name is extracted from the geographical names.
+     *
+     * @test
+     */
+    public function nameIsExtractedFromGeographicalNames(): void
+    {
+        $streetNameId = new StreetNameId(123);
+        $geographicalNames = new GeographicalNames(
+            new GeographicalName(new LanguageCode('NL'), 'Foo Nl')
+        );
+
+        $streetName = new StreetName($streetNameId, $geographicalNames);
+
+        $this->assertSame($geographicalNames->name(), $streetName->name());
     }
 
     /**
@@ -40,7 +58,7 @@ class StreetNameTest extends TestCase
      *
      * @test
      */
-    public function notSameIfObjectIdIsDifferent(): void
+    public function notSameIfStreetNameIdIsDifferent(): void
     {
         $geographicalNames = new GeographicalNames(
             new GeographicalName(new LanguageCode('EN'), 'Foo EN')
@@ -51,6 +69,28 @@ class StreetNameTest extends TestCase
 
         $otherStreetNameId = new StreetNameId(456);
         $otherStreetName = new StreetName($otherStreetNameId, $geographicalNames);
+
+        $this->assertFalse($streetName->sameValueAs($otherStreetName));
+    }
+
+    /**
+     * Not the same value if the geographical names are different.
+     *
+     * @test
+     */
+    public function notSameIfGeographicalNamesAreDifferent(): void
+    {
+        $geographicalNames = new GeographicalNames(
+            new GeographicalName(new LanguageCode('EN'), 'Foo EN')
+        );
+
+        $streetNameId = new StreetNameId(123);
+        $streetName = new StreetName($streetNameId, $geographicalNames);
+
+        $otherGeographicalNames = new GeographicalNames(
+            new GeographicalName(new LanguageCode('NL'), 'Foo NL')
+        );
+        $otherStreetName = new StreetName($streetNameId, $otherGeographicalNames);
 
         $this->assertFalse($streetName->sameValueAs($otherStreetName));
     }
@@ -71,5 +111,22 @@ class StreetNameTest extends TestCase
         $sameStreetName = new StreetName($streetNameId, $geographicalNames);
 
         $this->assertTrue($streetName->sameValueAs($sameStreetName));
+    }
+
+    /**
+     * Casting to string returns name from GeographicalNames collection.
+     *
+     * @test
+     */
+    public function castToStringReturnsPostalCodeAndName(): void
+    {
+        $streetNameId = new StreetNameId(123);
+        $geographicalNames = new GeographicalNames(
+            new GeographicalName(new LanguageCode('EN'), 'Foo EN')
+        );
+
+        $streetName = new StreetName($streetNameId, $geographicalNames);
+
+        $this->assertSame($geographicalNames->name(), (string) $streetName);
     }
 }