+++ CHANGELOG.md
@@ -13,5 +13,6 @@ All Notable changes to `digipolisgent/flanders-basicregisters` package.
 * Added the address detail.
 * Added the address matches related objects.
 * Added the post info and collection objects.
+* Added the Localities collection.
 
 [Unreleased]: https://github.com/digipolisgent/php_package_dg-flanders-basicregisters/compare/master...develop